+++ lib/load.php
@@ -107,6 +107,7 @@ require __DIR__ . '/compat/wordpress-7.1/icons.php';
 
 // WordPress 7.2 compat.
 require __DIR__ . '/compat/wordpress-7.2/kses.php';
+require __DIR__ . '/compat/wordpress-7.2/block-template-utils.php';
 
 // Experimental features.
 require __DIR__ . '/experimental/block-editor-settings-mobile.php';