+++ packages/editor/src/components/post-template/hooks.js
@@ -1,6 +1,6 @@
 import { useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
-import { useEntityProp, store as coreStore } from '@wordpress/core-data';
+import { store as coreStore } from '@wordpress/core-data';
 import { __, sprintf } from '@wordpress/i18n';
 import { store as editorStore } from '../../store';
 
@@ -15,111 +15,51 @@ export function useEditedPostContext() {
 }
 export function useAllowSwitchingTemplates() {
 	const { postType, postId } = useEditedPostContext();
-	return useSelect(
-		( select ) => {
-			const { canUser, getEntityRecord, getEntityRecords } =
-				select( coreStore );
-			const siteSettings = canUser( 'read', {
-				kind: 'root',
-				name: 'site',
-			} )
-				? getEntityRecord( 'root', 'site' )
-				: undefined;
-
-			const isPostsPage = +postId === siteSettings?.page_for_posts;
-			const isFrontPage =
-				postType === 'page' && +postId === siteSettings?.page_on_front;
-			// If current page is set front page or posts page, we also need
-			// to check if the current theme has a template for it. If not
-			const templates = isFrontPage
-				? getEntityRecords( 'postType', 'wp_template', {
-						per_page: -1,
-					} )
-				: [];
-			const hasFrontPage =
-				isFrontPage &&
-				!! templates?.some( ( { slug } ) => slug === 'front-page' );
-			return ! isPostsPage && ! hasFrontPage;
-		},
-		[ postId, postType ]
-	);
+	const templates = useTemplates( postType, postId );
+	return templates?.length > 1;
 }
 
-function useTemplates( postType ) {
+function useTemplates( postType, postId ) {
 	return useSelect(
 		( select ) =>
 			select( coreStore ).getEntityRecords( 'postType', 'wp_template', {
 				per_page: -1,
 				post_type: postType,
-				// We look at the combined templates for now (old endpoint)
-				// because posts only accept slugs for templates, not IDs.
+				post_id: Number( postId ),
 			} ),
-		[ postType ]
+		[ postType, postId ]
 	);
 }
 
 export function useAvailableTemplates() {
 	const { postType, postId } = useEditedPostContext();
-	const [ postSlug ] = useEntityProp( 'postType', postType, 'slug', postId );
 	const currentTemplateSlug = useCurrentTemplateSlug();
 	const allowSwitchingTemplate = useAllowSwitchingTemplates();
-	const templates = useTemplates( postType );
-	// Add the default template to the available ones. We don't care about
-	// possible assignment to postspage/homepage because it's guarded by
-	// `allowSwitchingTemplate` above.
-	const defaultTemplate = useSelect(
-		( select ) => {
-			// Only append the default template if the experiment is enabled.
-			if ( ! window?.__experimentalDataFormInspector ) {
-				return null;
-			}
-			// If the default template is already assigned, no need
-			// to add it to the available templates.
-			if ( ! currentTemplateSlug ) {
-				return null;
-			}
-			const { getDefaultTemplateId, getEntityRecord } =
-				select( coreStore );
-			let slug;
-			if ( postSlug ) {
-				slug =
-					postType === 'page'
-						? `${ postType }-${ postSlug }`
-						: `single-${ postType }-${ postSlug }`;
-			} else {
-				slug = postType === 'page' ? 'page' : `single-${ postType }`;
-			}
-			const templateId = getDefaultTemplateId( { slug } );
-			if ( ! templateId ) {
-				return null;
-			}
-			return getEntityRecord( 'postType', 'wp_template', templateId );
-		},
-		[ currentTemplateSlug, postSlug, postType ]
-	);
+	const templates = useTemplates( postType, postId );
+	const defaultTemplate = templates?.[ 0 ];
 	return useMemo(
 		() =>
 			allowSwitchingTemplate &&
 			[
 				...( templates || [] ).filter(
 					( template ) =>
-						template.is_custom &&
-						template.slug !== currentTemplateSlug &&
-						!! template.content.raw // Skip empty templates.
+						template.id !== defaultTemplate?.id &&
+						template.slug !== currentTemplateSlug
 				),
-				defaultTemplate && {
-					...defaultTemplate,
-					title: {
-						rendered: sprintf(
-							// translators: %s: Template name
-							__( '%s (default)' ),
-							defaultTemplate.title.rendered
-						),
+				currentTemplateSlug &&
+					defaultTemplate && {
+						...defaultTemplate,
+						title: {
+							rendered: sprintf(
+								// translators: %s: Template name
+								__( '%s (default)' ),
+								defaultTemplate.title.rendered
+							),
+						},
+						// That's extra custom prop in order to update to an empty template
+						// when we select the default template.
+						isDefault: true,
 					},
-					// That's extra custom prop in order to update to an empty template
-					// when we select the default template.
-					isDefault: true,
-				},
 			].filter( Boolean ),
 		[
 			templates,
@@ -174,9 +114,12 @@ export function usePostTemplatePanelMode() {
 
 export function useCurrentTemplateSlug() {
 	const { postType, postId } = useEditedPostContext();
-	const templates = useTemplates( postType );
+	const templates = useTemplates( postType, postId );
 	const entityTemplate = useSelect(
 		( select ) => {
+			if ( templates?.length === 1 ) {
+				return templates[ 0 ].slug;
+			}
 			const post = select( coreStore ).getEditedEntityRecord(
 				'postType',
 				postType,
@@ -184,15 +127,16 @@ export function useCurrentTemplateSlug() {
 			);
 			return post?.template;
 		},
-		[ postType, postId ]
+		[ postType, postId, templates ]
 	);
 
 	if ( ! entityTemplate ) {
 		return;
 	}
-	// If a page has a `template` set and is not included in the list
-	// of the theme's templates, do not return it, in order to resolve
-	// to the current theme's default template.
+	if ( templates?.length > 1 && entityTemplate === templates[ 0 ].slug ) {
+		return;
+	}
+	// An assignment outside the returned choices falls back to the default.
 	return templates?.find( ( template ) => template.slug === entityTemplate )
 		?.slug;
 }